audio-processor/requirements.txt
absl-py==1.0.0
accelerate==0.31.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.2.0
alabaster==0.7.8
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==3.6.1
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
arrow==0.15.5
astor==0.8.1
astroid==2.3.3
astunparse==1.6.3
async-timeout==5.0.1
atomicwrites==1.1.5
attrs==19.3.0
Automat==0.8.0
autopep8==1.5
Babel==2.6.0
backcall==0.1.0
bandit==1.6.2
base58==1.0.3
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
blis==0.4.1
bokeh==2.0.2
boto==2.49.0
boto3==1.23.6
botocore==1.26.6
bpemb==0.3.2
brewer2mpl==1.4.1
cachetools==4.1.0
catalogue==2.0.6
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
cloud-init==24.4.1
cloudpickle==1.3.0
cmake==3.26.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
cytoolz==0.10.1
dataclasses-json==0.6.7
datasets==2.20.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.9
dill==0.3.1.1
distlib==0.3.0
distro==1.9.0
distro-info==0.23+ubuntu1.1
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
en-core-web-lg==3.2.0
en-core-web-md==3.2.0
en-core-web-sm==3.2.0
entrypoints==0.3
et-xmlfile==1.0.1
etils==1.2.0
eval-type-backport==0.2.2
evaluate==0.4.0
fail2ban==0.11.1
faiss-cpu==1.8.0.post1
fastapi==0.124.4
fastjsonschema==2.16.2
fasttext==0.9.2
filelock==3.0.12
fire==0.2.1
flair==0.6.0.post1
flake8==3.7.9
flatbuffers==23.3.3
frozenlist==1.3.0
fsspec==2024.3.1
ftfy==5.8
future==0.18.2
gast==0.3.3
gensim==4.0.1
ggplot==0.11.5
gitdb2==2.0.6
GitPython==3.0.7
google-auth==1.14.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
greenlet==3.1.1
grpcio==1.62.0
h11==0.14.0
h5py==2.10.0
html5lib==1.0.1
httpcore==1.0.7
httplib2==0.14.0
httpx==0.28.1
huggingface-hub==0.23.4
hydra-core==1.0.6
hyperlink==19.0.0
hyperopt==0.2.4
idna==2.8
imagesize==1.2.0
importlib-metadata==5.0.0
importlib-resources==5.1.2
incremental==16.10.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==4.3.4
Janome==0.4.0
jax==0.4.8
jdcal==1.0
jedi==0.15.2
jellyfish==0.7.2
Jinja2==3.1.2
jiter==0.9.0
jmespath==0.9.5
joblib==1.4.2
json5==0.9.4
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.5
jupyter-console==6.0.0
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.2
jupyterlab-widgets==3.0.3
kenlm==0.0.0
keras==2.13.1
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
konoha==4.6.1
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langcodes==3.3.0
langdetect==1.0.8
langsmith==0.1.147
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libclang==14.0.1
lit==16.0.1
logilab-common==1.4.3
lxml==4.5.0
Markdown==3.2.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.22.0
matplotlib==3.1.2
mccabe==0.6.1
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.1.0
mock==3.0.5
more-itertools==9.1.0
mpi4py==3.0.3
mpld3==0.3
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.2
mypy==0.761
mypy-extensions==0.4.3
mysqlclient==1.4.4
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
notebook-shim==0.1.0
numexpr==2.8.4
numpy==1.23.5
numpydoc==0.7.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.6
openai==1.66.3
OpenCC==1.1.9
openpyxl==3.0.3
opt-einsum==3.3.0
orjson==3.10.15
overrides==3.0.0
packaging==24.2
pandas==2.0.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.2
pathtools==0.1.2
pathy==0.6.1
patsy==0.5.1
pbr==5.4.5
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
pipenv==11.9.0
plac==0.9.6
platformdirs==4.2.2
plotly==5.14.1
pluggy==0.13.0
portalocker==2.4.0
preshed==3.0.2
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
propcache==0.2.0
protobuf==3.20.3
psutil==5.5.1
ptyprocess==0.7.0
py==1.8.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.5.0
pycodestyle==2.5.0
pycparser==2.21
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.8.1b0
PyDispatcher==2.0.5
pyemd==0.5.1
pyenchant==3.1.1
pyflakes==2.1.1
Pygments==2.13.0
PyGObject==3.36.0
pygtrie==2.4.2
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.4.4
pylint-common==0.2.5
pylint-plugin-utils==0.6
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
Pyphen==0.9.5
PyQt5==5.14.1
pyrsistent==0.15.5
pyserial==3.4
pysvn==1.9.9
pytest==5.4.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-Levenshtein==0.12.2
python-multipart==0.0.20
pytz==2023.3
PyYAML==5.3.1
pyzmq==24.0.1
QtAwesome==0.4.4
qtconsole==4.6.0
QtPy==1.9.0
queuelib==1.5.0
redis==6.1.1
regex==2023.3.23
requests==2.32.3
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
responses==0.18.0
retrying==1.3.3
rich==13.3.5
roman==2.0.0
rope==0.10.5
rq==2.3.3
rsa==4.0
s3transfer==0.5.2
sacrebleu==2.1.0
sacremoses==0.0.41
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
Scrapy==1.7.3
screen-resolution-extra==0.0.0
seaborn==0.10.0
SecretStorage==2.3.1
segtok==1.5.9
Send2Trash==1.8.0
sentencepiece==0.1.98
sentry-sdk==2.22.0
seqeval==1.2.2
service-identity==18.1.0
setproctitle==1.3.2
simplejson==3.16.0
sip==4.19.21
six==1.15.0
sklearn-pandas==1.8.0
smart-open==5.2.1
smmap2==2.0.5
sniffio==1.3.0
sos==4.8.2
soupsieve==1.9.5
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
Sphinx==1.8.5
spyder==3.3.6
spyder-kernels==0.5.2
sqlalchemy==2.0.40
sqlitedict==1.6.0
srsly==2.4.2
ssh-import-id==5.10
starlette==0.44.0
statsmodels==0.11.1
stevedore==1.32.0
streamlit==1.40.1
streamlit-aggrid==0.3.4.post3
sympy==1.11.1
systemd-python==234
tables==3.6.1
tabulate==0.8.9
tenacity==8.2.2
tensorboard==2.14.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.13.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.13.1
termcolor==1.1.0
terminado==0.16.0
testpath==0.4.4
textacy==0.10.0
thinc==8.0.13
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.1.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.9.0
torch==2.2.1
torchaudio==2.2.1
torchdata==0.6.1
torchtext==0.15.2+cpu
torchvision==0.17.1
tornado==6.2
tqdm==4.66.4
traitlets==5.4.0
transformers==4.41.2
triton==2.2.0
Twisted==18.9.0
typed-ast==1.4.1
typer==0.4.0
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==2.1b1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.26.15
uvicorn==0.33.0
validators==0.20.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
w3lib==1.21.0
wadllib==1.3.3
wandb==0.19.8
wasabi==0.9.0
watchdog==3.0.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==1.0.1
widgetsnbextension==4.0.3
wrapt==1.12.1
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
xxhash==3.0.0
yapf==0.29.0
yarl==1.15.2
zipp==1.0.0
zope.interface==4.7.1
